perf(re-ordering of components): added useHoverComponent hooks
useHoverComponent hooks are added to handle the ordering of components.\
