+++ src/components/editor/ComponentPreview.tsx
@@ -34,7 +34,6 @@ const ComponentPreview: React.FC<{
   }
 
   const type = (component && component.type) || null
-  console.log(type)
   switch (type) {
     // Simple components
     case 'Badge':
