+++ src/hooks/useDropComponent.ts
@@ -1,54 +1,16 @@
 import { useDrop, DropTargetMonitor } from 'react-dnd'
-import { useSelector } from 'react-redux'
 import { rootComponents } from '../utils/editor'
 import useDispatch from './useDispatch'
 import builder from '../core/models/composer/builder'
-import { getComponents } from '../core/selectors/components'
 
 export const useDropComponent = (
   componentId: string,
   accept: (ComponentType | MetaComponentType)[] = rootComponents,
   canDrop: boolean = true,
-  boundingPosition?: {
-    top: number
-    bottom: number
-  },
 ) => {
   const dispatch = useDispatch()
-  const components = useSelector(getComponents)
   const [{ isOver }, drop] = useDrop({
     accept,
-    hover: (item, monitor) => {
-      if (item.isMoved && boundingPosition) {
-        if (componentId === item.id) {
-          return
-        }
-        const selectedComponent = components[item.id]
-        const { top, bottom } = boundingPosition
-        const hoverMiddleY = (bottom - top) / 2
-        const clientOffset = monitor.getClientOffset()
-        const fromIndex = components[selectedComponent.parent].children.indexOf(
-          item.id,
-        )
-        const toIndex = components[selectedComponent.parent].children.indexOf(
-          componentId,
-        )
-        const hoverClientY = clientOffset && clientOffset.y - top
-
-        // hasPassedMid
-        if (fromIndex < toIndex && hoverClientY && hoverClientY < hoverMiddleY)
-          return
-
-        if (fromIndex > toIndex && hoverClientY && hoverClientY > hoverMiddleY)
-          return
-
-        dispatch.components.moveSelectedComponentChildren({
-          parentId: selectedComponent.parent,
-          fromIndex,
-          toIndex: toIndex === -1 ? fromIndex : toIndex,
-        })
-      }
-    },
     collect: monitor => ({
       isOver: monitor.isOver({ shallow: true }) && monitor.canDrop(),
     }),
