+++ src/hooks/useInteractive.ts
@@ -7,11 +7,11 @@ import {
   getIsHovered,
 } from '../core/selectors/components'
 import { getShowLayout, getFocusedComponent } from '../core/selectors/app'
-import { useDropComponent } from './useDropComponent'
 
 export const useInteractive = (
   component: IComponent,
   enableVisualHelper: boolean = false,
+  drop?: Function,
 ) => {
   const dispatch = useDispatch()
   const showLayout = useSelector(getShowLayout)
@@ -24,18 +24,9 @@ export const useInteractive = (
   })
 
   const ref = useRef<HTMLDivElement>(null)
-  let boundingPosition
-  boundingPosition =
+  const boundingPosition =
     ref.current !== null ? ref.current.getBoundingClientRect() : undefined
-  const { drop } = useDropComponent(
-    component.id,
-    undefined,
-    undefined,
-    boundingPosition && {
-      top: boundingPosition.top,
-      bottom: boundingPosition.bottom,
-    },
-  )
+
   let props = {
     ...component.props,
     onMouseOver: (event: MouseEvent) => {
@@ -74,5 +65,10 @@ export const useInteractive = (
     }
   }
 
-  return { props, ref: drag(drop(ref)), drag }
+  return {
+    props,
+    ref: drop ? drag(drop(ref)) : drag(ref),
+    drag,
+    boundingPosition,
+  }
 }
